+++ tests/hpke/mod.rs
@@ -1,0 +1,3 @@
+#[cfg(feature = "safe_api")]
+#[cfg(test)]
+pub mod hpke_test_kats;